// yumfe/src/app/remote/model/DailyOrder.ts
/**
 * Yum Food Orders
 * **Yum application, order food daily from the best chef in town**  This API is used by the angular.io client, and is not meant to be used otherwise.  Find source code of this API [here](https://github.com/jrtdev/yum)  Copyright (C) 2017 JR Technologies.     ------------------------------------      Last edit: 28/09/2017 10:00  ------------------------------------- 
 *
 * OpenAPI spec version: 1.2.8
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

/**
 * A single order DTO
 */
export interface DailyOrder {
    foods?: Array<models.FoodWithQuantity>;

    dailyMenuDate?: Date;

    comment?: string;

    lastEdit?: models.LastEdit;

}